src/streamlit/requirements.txt
pandas=1.1.0
plotly
streamlit
seaborn
tensorflow-cpu==2.3.0